Implement new logging/printing standard for BoomerAMG.  This is a minimal implementation needed to support the Babel interface.  The changes are:
 Eliminate the BoomerAMGSetLogging functions, and the 3-argument
BoomerAMGSetPrintLevel functions.  In their place are 2-argument
BoomerAMGSetPrintLevel and BoomerAMGSetPrintFileName functions.
Before and now, these serve to control how much (if any) data is printed,
and (in principle - never implemented) where it is printed to.

 Add BoomerAMGSetLogLevel functions.  Setting the log level does not control
printing, but does control additional computations of diagnostic information.
This involves an additional member, int log_level, of the struct ParAMGData.

 Add capability to store the last residual in an additional array.  The user
can get it through HYPRE_BoomerAMGGetResidual.  This is done only if
HYPRE_BoomerAMGSetLogLevel has been called, with a log level >2.  This involves
an additional member, double *residual (default NULL), of the struct ParAMGData.