1.  New variable owns_matvec_data to keep track of who created matvec_data (in struct hypre_PCGData).  It is possible for the user to "own" it, in which case we shouldn't delete it.
2.  In pcg.c, hypre_PCGSolve, there has long been a test on whether
sdotp==0 near the top of the main loop.  If the test fails, it breaks
out of the loop and the function returns after a brief calculation of
the residulal norm.  There is a case in which there may be invalid
data in i_prod, used for the norm calculation, possibley leading to a
crash.  So now i_prod is set to a valid (if meaningless) number in
that case.