Added a more robust version of AMS for problems with zero-conductivity regions, where the matrix is singular Ker(A)=Ran(G0).
To use this version, provide a list of the interior nodes in the zero
conductivity region using AMSSetInteriorNodes(). This will constructs G0 and AMS
for A+G0G0^T.  For compatibility, the PCG residual is periodically projected
onto Ker(G0^T).  This is done in AMSSolve() and can be adjusted with
AMSSetProjectionFrequency().