Changed _hypre_utilities.h to auto generate and changed some nvcc configure stuff
The nvcc changes will need to change again once the other GPU work is merged.
The new 'Makefile.nvcc' and 'Makefile.empty' were used for portability, since
'if/else' is not standard in make.  The only other time we have excluded files
from being compiled in hypre, we excluded entire directories, but the mechanism
was similar.  This may need to be tweaked a bit.

The _hypre_utilities.h file is now generated with the 'headers' script and
should not be modified directly.