Put GPU header files into _hypre_utilities.h and made other cosmetic changes
This has not been tested with nvcc.  Also, some of the routines in gpukernels.cu
are needed in utilities, so maybe this file needs to be split or relocated to
utilities.  The utilities files should not include from other directories in
hypre and relative includes (i.e., '../directory') are to be avoided.