Removed ComplexInnerProd and changed return type on InnerProd
Also changed PCG back to its original form because it should be correct as is.
The inner product routine does need to return Complex in general, but in PCG all
of the calls should return real values as long as the problem and preconditioner
are Hermitian.  An explicit cast from Complex to Real doesn't seem to be needed
for the GNU compiler (no warnings are generated); it takes the real part.

The tests run cleanly without complex turned on, but configuring with maxdim=4
and complex has some compile problems that need to be sorted out.