This version passes most of the regression tests
The C++ compile creates some errors as do some of the examples.  All of them are
related to blas and lapack code, which changed quite a bit.  The errors are
minor and should be straightforward to fix.  The standard runtests all run and
valgrind is clean through most of them (they are running now).

It should be okay to use this version to compare with Chris' original version in
bitbucket to make sure the results are the same.