I believe that GMRES relaxation is now working with SysBAMG solver.  Although, there's still a regression test issue to work out.  I had to add a new set of Krylov wrappers for SStructPVector and SStructPMatrix data types, because that is the data type used for computation inside of the SysBAMG solver.  So this required some new files like krylov_sstructp.c and hypre_sstructp_gmres.c
Because using GMRES requires pre-allocating a Krylov workspace, and I want to
keep the solver code from getting too complex, I decided to set the number of
pre and post relaxation sweeps during the BAMG setup phase, whereas they used
to be set on-the-fly during the solve phase.

I also now enforce the number
of pre-relaxations to equal the number of post-relaxations so that the Krylov
workspace is the right size for both pre- and post-relaxation.

Added an ability to run unpreconditioned BAMG in the sstruct driver.
Previously, you could only run BAMG with CG as a preconditioner, but that
doesn't allow for experimentation with a GMRES smoother.

Changed dimensions in call to hypre_xxxmqr() in sys_bamg_setup_interp.c
Valgrind was giving and error, and I believe this to be correct