Added a HYPRE_BuildInfo() function
It is set up in the 'lib' directory so that the build information is created
each time a new library file is created.  The script 'buildinfo.sh' creates the
information by calling various utilities and mimicking a C and C++ compile.

The 'sstruct.c' driver has a -buildinfo option for testing the new feature.

This is not yet set up to work with CMake.  Some of the same files and scripts
could be used there, but only on a linux/unix system.  I am not sure how to do
this more generally (yet).