Update unified-memory testing logic to account for HIP
There's no reason to look for CUDA if we're using HIP with unified
memory. So update the logic to test for CUDA if using CUDA or
OpenMP with device  and test for HIP if using HIP. And then error
if unified memory was requested but none of those options were
also requested.