Remove hypre_CSRMatrixMatvecMaskedDevice()
In cuda 11, the bsrxmv masked matvec routine provided by cusparse no
longer supports blockDim = 1, creating an error in the masked matvec
implementation in hypre. The only place this is currently used is in
the AMG-DD CF L1 Jacobi smoother. Therefore, I simply removed the
masked matvec routine and replaced it with a regular matvec inside
the smoother.