Implement one point interpolation
This is a first draft implementation of one point interpolation
on the device. I also continue debugging the NeumannAIR restriction
operator. This is looking very close now, but I get different
resulting restriction operators on coarser grids (finest grid looks
the same) on the CPU vs. the GPU. I think this may be due to the
fact that the strength thresholding for R is done differently on
the CPU vs. GPU. Further investigation in future commits.