Add CF relaxation on the device for L1 Jacobi
This is an easy implementation that does some redundant computation.
Namely, this performs a full matrix-vector multiply, but then only
updates either C or F points as requested. Some changes in the
underlying device routines led to some restructuring and simplification
of the AMG-DD code where this CF strategy was used previously. This
still needs testing.