Change names and fix initialization
This does a bunch of name changing of files, data structures,
and variables from 'cuda' to 'device' in order to reflect which
things are generic device functionality vs. tied to a specific
language. In addition, this now compiles and runs a simple program
that calls HYPRE_Init() and allocates/copies/frees memory on the
device and host with unified memory.