Fix shared-libraries build for HIP
We're not sucking in CXXFLAGS because if the user enables --with-openmp,
this mucks up the compilation (with or without shared libs) because
the HIP compiler is clang under-the-hood and the clang openmp host
compilation pass does not play nice with the GCC openmp compilation
pass at link time. So instead we just manually enable the shared lib
flags for the HIP compiler.