Silence uninitialized var HIP warnings
This should be a benign change. What happens is that the first one or two
workitems/threads in each workgroup/block read a value and then broadcast
it (with __shfl_sync or similar) and then code branching happens
based on this value. But the compiler can't see it all the way
through it so we get some uninitialized var warnings.