Additional options for Chebyshev-accelerated Jacobi smoothing.
Variant 2: Standard, 1st Kind Chebyshev smoothing with $D^{-1}A$.
No restriction on the Chebyshev order.

Variant 3: 4th Kind Chebyshev smoother with $D^{-1}A$.
No restriction on the Chebyshev order.

Variant 4: Optimized 4th Kind Chebyshev smoother with $D^{-1}A$.
Chebyshev order must be in [1,16] due to needing to compute the
$\beta_k$ coefficients.

Variants 3-4 are from https://arxiv.org/pdf/2202.08830.pdf.

These correspond to relaxation type 16.

In addition, relaxation type 19 has been implemented as a "no-op"
smoother. This can be used in the case a user wishes to omit the post
smoothing step, for example.